50 percent enrollment modifications
Check that a student was enrolled 50 percent or more of the school year
at a school before taking maximum number of ISP days. This ensure we
don't identify an accountability system/school that has more ISP days
but represents an enrollment of less than 50 percent. Also remove
enrollments in CTE, Alternative, Adult HS.